refactor: move updaters
